Arreglo el procesado.log del consolidar.sh

+++ consolidar.sh
@@ -5,10 +5,13 @@ while true
 do 
     for archivo in $HOME/EPNro1/entrada/*.txt
     do 
-        FECHA_HORA=$(date "+%d/%m/%Y %T")
-        cat "$archivo" >> $HOME/EPNro1/salida/$FILENAME.txt
-        mv "$archivo" $HOME/EPNro1/procesado/
-        echo "$FECHA_HORA - Procesado archivo $(basename "$archivo")" >> $HOME/EPNro1/procesado/procesado.log
+        if [ -f "$archivo" ];
+        then
+            FECHA_HORA=$(date "+%d/%m/%Y %T")
+            cat "$archivo" >> $HOME/EPNro1/salida/$FILENAME.txt
+            mv "$archivo" $HOME/EPNro1/procesado/
+            echo "$FECHA_HORA - Procesado archivo $(basename "$archivo")" >> $HOME/EPNro1/procesado/procesado.log
+        fi
     done
 sleep 30    
 done
